Improved disease genes predictions
The training matrix was not correctly ensembled. As the positive samples were being concatenated, all of them were found at the end of the matrix. Thus, if the cross validation, when the split test corresponded to the last part (the matrix ending), the performance drop considerably and this was because the training set had all or most of them negative samples. This was corrected by shuffeling the samples (with `training_matrix = training_matrix.sample(frac=1)` )